requirements.txt
pyjwt==1.7.1
scrypt==0.8.18
flask==1.1.4
flask-SQLAlchemy==2.5.1
flask_script==2.0.6
jinja2==2.10.1
markupsafe==2.0.1
gunicorn==20.0.4
pytest==7.1.2
flasgger==0.9.5